+++ src/networks/types/ERE.js
@@ -5,9 +5,9 @@ import ere from '@/assets/images/networks/ere.svg';
 export default {
   name: 'ERE',
   name_long: 'EtherCore',
-  homePage: 'https://ethercore.org',
-  blockExplorerTX: 'https://explorer.ethercore.org/tx/[[txHash]]',
-  blockExplorerAddr: 'https://explorer.ethercore.org/address/[[address]]',
+  homePage: 'https://ethercore.io',
+  blockExplorerTX: 'https://explorer.ethercore.io/tx/[[txHash]]',
+  blockExplorerAddr: 'https://explorer.ethercore.io/address/[[address]]',
   chainID: 466,
   tokens: tokens,
   contracts: contracts,