+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### Feature
+
+- Unstoppable domain buy domain [#2229] (https://github.com/MyEtherWallet/MyEtherWallet/pull/2229)
+
 ### Devop
 
 - Update packages [#2254](https://github.com/MyEtherWallet/MyEtherWallet/pull/2254)