+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+### Release v5.5.4
+
 ### Feature
 
 - ENS Avatar [#2238](https://github.com/MyEtherWallet/MyEtherWallet/pull/2238)
@@ -5,6 +7,8 @@
 
 ### Devop
 
+- Update ATH node [#2189](https://github.com/MyEtherWallet/MyEtherWallet/pull/2189)
+- Update EthCore node [#2251](https://github.com/MyEtherWallet/MyEtherWallet/pull/2251)
 - Update packages [#2254](https://github.com/MyEtherWallet/MyEtherWallet/pull/2254)
 - Cx workflow [#2237](https://github.com/MyEtherWallet/MyEtherWallet/pull/2237)